extract likecount to helper, session storage broken
